[refactor] data 패키지 구조 변경
data 패키지 복구
