+++ data/src/androidTest/java/com/daejol/data/FakeRepositoryModule.kt
@@ -1,12 +1,12 @@
 package com.daejol.data
 
-import com.daejol.catdata.repository.CatImagesRepositoryImpl
+import com.daejol.data.catdata.repository.CatImagesRepositoryImpl
 import com.daejol.domain.repository.CatImagesRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.components.SingletonComponent
 import dagger.hilt.testing.TestInstallIn
-import di.RepositoryModule
+import com.daejol.data.di.RepositoryModule
 import javax.inject.Singleton
 
 @Module
