+++ data/src/androidTest/java/com/daejol/data/RetrofitTest.kt
@@ -1,7 +1,7 @@
 package com.daejol.data
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.daejol.catdata.api.CatImagesApi
+import com.daejol.data.catdata.api.CatImagesApi
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.After
 import org.junit.Before
