+++ data/src/main/java/com/daejol/data/di/AppModule.kt
@@ -1,11 +1,10 @@
-package di
+package com.daejol.data.di
 
-import com.daejol.catdata.api.CatBreedsApi
-import common.ApiConst
-import com.daejol.catdata.api.CatImagesApi
+import com.daejol.data.catdata.api.CatBreedsApi
+import com.daejol.data.ApiConst
+import com.daejol.data.catdata.api.CatImagesApi
 import com.daejol.catdata.api.DogImagesApi
-import com.daejol.catdata.dto.CatBreedsDto
-import com.daejol.dogdata.api.DogBreedsApi
+import com.daejol.data.dogdata.api.DogBreedsApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
