+++ data/src/main/java/com/daejol/data/dogdata/dto/DogBreedDto.kt
@@ -1,4 +1,4 @@
-package com.daejol.dogdata.dto
+package com.daejol.data.dogdata.dto
 
 import com.google.gson.annotations.SerializedName
