+++ data/src/main/java/com/daejol/data/dogdata/dto/DogImageDto.kt
