+++ data/src/main/java/com/daejol/data/dogdata/repository/DogBreedsRepositoryImpl.kt
@@ -1,9 +1,9 @@
-package com.daejol.dogdata.repository
+package com.daejol.data.dogdata.repository
 
 import DataState
 import com.daejol.catdata.dto.DogMapper.toDomain
 import com.daejol.domain.repository.DogBreedsRepository
-import com.daejol.dogdata.api.DogBreedsApi
+import com.daejol.data.dogdata.api.DogBreedsApi
 import entity.BreedInfoEntity
 import entity.BreedTypeEntity
 import kotlinx.coroutines.flow.Flow
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
 class DogBreedsRepositoryImpl @Inject constructor(
-     private val dogBreedsApi: DogBreedsApi,
+    private val dogBreedsApi: DogBreedsApi,
 ): DogBreedsRepository {
 
     override suspend fun getDogBreeds(): Flow<DataState<List<BreedInfoEntity>>> = flow {
